+++ fchmod.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021-2022 Gaël PORTAY
+ * Copyright 2021-2023 Gaël PORTAY
  *
  * SPDX-License-Identifier: LGPL-2.1-or-later
  */
@@ -16,7 +16,8 @@ int fchmod(int fd, mode_t mode)
 {
 	int atflags = AT_EMPTY_PATH;
 
-	__debug("%s(fd: %i, mode: 0%03o)\n", __func__, fd, mode);
+	__debug("%s(fd: %i <-> '%s', mode: 0%03o)\n", __func__, fd,
+		__fpath(fd), mode);
 
 #ifdef __linux__
 	atflags = 0;