+++ fchown.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021-2022 Gaël PORTAY
+ * Copyright 2021-2023 Gaël PORTAY
  *
  * SPDX-License-Identifier: LGPL-2.1-or-later
  */
@@ -14,8 +14,8 @@
 
 int fchown(int fd, uid_t owner, gid_t group)
 {
-	__debug("%s(fd: %i, owner: %i, group: %i)\n", __func__, fd, owner,
-		group);
+	__debug("%s(fd: %i <-> '%s', owner: %i, group: %i)\n", __func__, fd,
+		__fpath(fd), owner, group);
 
 	return fchownat(fd, "", owner, group, AT_EMPTY_PATH);
 }