+++ fexecve.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021-2022 Gaël PORTAY
+ * Copyright 2021-2023 Gaël PORTAY
  *
  * SPDX-License-Identifier: LGPL-2.1-or-later
  */
@@ -37,8 +37,8 @@ static int __fexecve(int fd, char * const argv[], char * const envp[])
 
 int fexecve(int fd, char * const argv[], char * const envp[])
 {
-	__debug("%s(fd: %i, argv: { '%s', '%s', ... }, envp: %p)\n",
-		__func__, fd, argv[0], argv[1], envp);
+	__debug("%s(fd: %i <-> '%s', argv: { '%s', '%s', ... }, envp: %p)\n",
+		__func__, fd, __fpath(fd), argv[0], argv[1], envp);
 
 	return __fexecve(fd, argv, envp);
 }