+++ fstat.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021-2022 Gaël PORTAY
+ * Copyright 2021-2023 Gaël PORTAY
  *
  * SPDX-License-Identifier: LGPL-2.1-or-later
  */
@@ -14,7 +14,7 @@
 
 int fstat(int fd, struct stat *statbuf)
 {
-	__debug("%s(fd: %i, ...)\n", __func__, fd);
+	__debug("%s(fd: %i <-> '%s', ...)\n", __func__, fd, __fpath(fd));
 
 	return fstatat(fd, "", statbuf, AT_EMPTY_PATH);
 }